Marking To Dos to Start Implementing New Chart Library
I used comments to mark the parts of the code that I think need to be added/deleted/altered for us to switch from the ngx-chart library to the apex-chart library.

We should try implementing the new library and ensure that it works completely before we delete any of the old chart library code.